SmartGrid_Project_Devops/docker/api_gateway/requirements.txt
flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-flask==0.42b0
opentelemetry-instrumentation-requests==0.42b0
opentelemetry-exporter-jaeger-thrift==1.21.0
redis==5.0.1
